Fix Some Build Mistakes and Update Docs

+++ task-1/include/playground/utils.hpp
@@ -40,7 +40,7 @@ void initRandMat(std::size_t m, std::size_t n, T* mat)
 {
     srand(time(nullptr));
     for (std::size_t cnt = 0; cnt < m * n; cnt++) {
-        mat[cnt] = T(float32_t(rand()) / RAND_MAX +
+        mat[cnt] = T(float32_t(rand()) / float32_t(RAND_MAX) +
                      std::numeric_limits<float32_t>::min());
     }
 }
